Executor: Generate AIR proof input
